dev_con/dev_cont/pyproject.toml
[tool.poetry]
name = "dev-con"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pytest = "^8.1.1"
httpx = "^0.27.0"
sqlmodel = "^0.0.18"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.2.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
